+++ test/utils.py
@@ -29,6 +29,7 @@ from functools import partial
 
 from pymongo import MongoClient, monitoring
 from pymongo.errors import AutoReconnect, OperationFailure
+from pymongo.monitoring import CommandStartedEvent
 from pymongo.server_selectors import (any_server_selector,
                                       writable_server_selector)
 from pymongo.write_concern import WriteConcern
@@ -62,10 +63,17 @@ class WhiteListEventListener(monitoring.CommandListener):
 
 class EventListener(monitoring.CommandListener):
 
-    def __init__(self):
+    def __init__(self, ignore_lsid=True):
+        self.ignore_lsid = ignore_lsid
         self.results = defaultdict(list)
 
     def started(self, event):
+        if self.ignore_lsid and 'lsid' in event.command:
+            cmd = event.command.copy()
+            cmd.pop('lsid', None)
+            event = CommandStartedEvent(cmd, event.database_name,
+                                        event.request_id, event.connection_id,
+                                        event.operation_id)
         self.results['started'].append(event)
 
     def succeeded(self, event):